refatornado getAll e colocando todos os arquivos em um so arquivo
