--- src/ListaDeTarefas.Application/UseCases/TarefaUseCase/GetAll/GetAllTarefaMapper.cs
@@ -1,13 +1,0 @@
-﻿using AutoMapper;
-using ListaDeTarefas.Domain.Entities;
-
-namespace ListaDeTarefas.Application.UseCases.TarefaUseCase.GetAll
-{
-    public sealed class GetAllTarefaMapper : Profile
-    {
-        public GetAllTarefaMapper() 
-        {
-            CreateMap<Tarefa, GetAllTarefaResponse>();
-        }
-    }
-}
