--- src/ListaDeTarefas.Application/UseCases/TarefaUseCase/GetAll/GetAllTarefaRequest.cs
@@ -1,8 +1,0 @@
-﻿using MediatR;
-
-namespace ListaDeTarefas.Application.UseCases.TarefaUseCase.GetAll
-{
-    public sealed record class GetAllTarefaRequest : IRequest<List<GetAllTarefaResponse>>
-    {
-    }
-}
