--- src/ListaDeTarefas.Application/UseCases/TarefaUseCase/GetAll/GetAllTarefaValidator.cs
@@ -1,9 +1,0 @@
-﻿using FluentValidation;
-
-namespace ListaDeTarefas.Application.UseCases.TarefaUseCase.GetAll
-{
-    public class GetAllTarefaValidator : AbstractValidator<GetAllTarefaRequest>
-    {
-        public GetAllTarefaValidator() { }
-    }
-}
